+++ test/week_calendar_test.exs
@@ -43,11 +43,11 @@ defmodule Cldr.Calendar.Week.Test do
   end
 
   test "day of week for ISOWeek calendar is correct" do
-    assert Cldr.Calendar.day_of_week(~d[2019-01-01]ISOWeek) == 1
+    assert Cldr.Calendar.day_of_week(~d[2019-01-01 ISOWeek]) == 1
   end
 
   test "day of week for NRF calendar is correct" do
-    assert Cldr.Calendar.day_of_week(~d[2019-01-01]NRF) == 7
+    assert Cldr.Calendar.day_of_week(~d[2019-01-01 NRF]) == 7
   end
 
   test "day of week for Sunday calendar is correct" do