+++ CHANGELOG.md
@@ -21,6 +21,8 @@ iex> IEx.configure(inspect: [inspect_fun: &Cldr.Calendar.inspect/2])
 
 * Ensure that `Cldr.Calendar` callbacks return a `Calendar.ISO` calendar if called with one (either as part of a date or as a separate argument).
 
+* Ensure the return calendar types of a an Interval are `Calendar.ISO` is the date provided is `Calendar.ISO`
+
 # Changelog for Cldr Calendars v0.9.0
 
 This is the changelog for Cldr v0.9.0 released on June 9th, 2019.  For older changelogs please consult the release tag on [GitHub](https://github.com/kipcole9/cldr_calendars/tags)